Refactored cover art file tests into format specific test files
