+++ test/test_coverart_image.py
@@ -1,6 +1,9 @@
 #!/usr/bin/env python
 # coding: utf-8
+import os.path
 import unittest
+
+from PyQt5 import QtCore
 
 from test.picardtestcase import (
     PicardTestCase,
@@ -83,6 +86,27 @@ class CoverArtImageTest(PicardTestCase):
         self.assertNotEqual(image1, image2)
         self.assertEqual(image2, image3)
         self.assertNotEqual(image2, image4)
+
+    def test_set_data(self):
+        self.addCleanup(QtCore.QObject.tagger.run_cleanup)
+        imgdata = data=create_fake_png(b'a')
+        imgdata2 = data=create_fake_png(b'xxx')
+        # set data once
+        coverartimage = CoverArtImage(data=imgdata2)
+        tmp_file = coverartimage.tempfile_filename
+        filesize = os.path.getsize(tmp_file)
+        # ensure file was written, and check its length
+        self.assertEqual(filesize, len(imgdata2))
+        self.assertEqual(coverartimage.data, imgdata2)
+
+        # set data again, with another payload
+        coverartimage.set_data(imgdata)
+
+        tmp_file = coverartimage.tempfile_filename
+        filesize = os.path.getsize(tmp_file)
+        # check file length again
+        self.assertEqual(filesize, len(imgdata))
+        self.assertEqual(coverartimage.data, imgdata)
 
 
 class LocalFileCoverArtImageTest(PicardTestCase):
